Refactored frontend code to support different evidence panes
