+++ public/src/components/TaskEvidencePane.jsx
@@ -1,0 +1,18 @@
+import React from "react";
+
+import { EvidencePaneProps } from "../types";
+
+/**
+ * @param {EvidencePaneProps} props The object containing the props.
+ * @returns {React.JSX.Element} The rendered pane.
+ */
+const TaskEvidencePane = ({
+  evidence,
+  containerModalId,
+  updateEvidenceFunc,
+}) => {
+  // TODO: the task evidence should be user stories
+  return <div></div>;
+};
+
+export default TaskEvidencePane;
